rake spec:integration uses nats 1.3.0 to run tests
when we moved to nats v2 we didn't update the integration test sandbox
to move away from the old gnatsd fork. This is the initial work to make
the sandbox use nats v2.

update gnatsd manager to install nats-server based on the binary we find
in the bosh-release blobs.

update nats.conf.erb fixture to work with natsv2

currently all tests that require that an agent process succesfully
connects to nats fail. this is because the certs generated through the
director are currently not synced into the nats auth json..