Short lived credentials for NATs as an opt in feature
Now we are not checking for the stemcell version to enable/disable the
NATS credentials rotation, but we rely on a configurable flag that will
rotate the credentials for all the VMs. Operators will have to make sure
that the stemcells they are using are compatible with this feature, or
else the incompatible stemcells will be unresponsive during a deploy.

[#183316690] [Deadline: Before End of FYQ1] [AHA] Short-lived NATS bootstrap credential